add an integration test for #32
